comments + more
I put comments for each method.

I noticed that my binary tree is a little off. Since I quick sorted the
array and then placed them in the binary tree in that order, I believe
my tree is much higher than I need it to be. However, this does not
matter so much because I end up searching the whole tree in getNearest
anyways. I did not have time to come up with a faster way of getting the
nearest using the binary tree.